chore(cleanup): create barrel files (#1192)
Creating barrel files for any module which exports anything other than
the default. This helps with controlling what is being shared and avoid
reaching between modules.


[Asana](https://app.asana.com/1/47184964732898/project/1204008699308084/task/1210525847101754?focus=true)
