+++ src/core/components/hyper-ref/index.ts
@@ -1,0 +1,2 @@
+export { default } from './hyper-ref';
+export { addHref } from './utils';
