+++ src/core/components/load-error/index.ts
@@ -1,0 +1,2 @@
+export { default } from './load-error';
+export type { Props } from './types';
