+++ src/core/components/loading/index.ts
@@ -1,0 +1,2 @@
+export { default } from './loading';
+export type { Props } from './types';
