+++ src/core/components/modal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './modal';
+export type { Props } from './types';
