+++ src/elements/hv-doc/index.ts
@@ -1,0 +1,2 @@
+export { default } from './hv-doc';
+export { StateContext } from './context';
