+++ src/elements/hv-route/index.ts
@@ -1,0 +1,2 @@
+export { default } from './hv-route';
+export type { Props } from './types';
